게시글 List 형태로 다중 삭제 처리 구현

+++ src/main/java/com/example/borderProject/spring/repository/BoardRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface BoardRepository extends JpaRepository<Board, Long> {
     public List<Board> findAll();
 
-    public Board findByBoardId(int boardId);
+    public Board findByBoardId(Long boardId);
 
     public Board deleteByBoardId(Board board);
 }
